correct pod log styling
